Modified project 01 and 03; Added data seeding and refactoring

+++ DB-with-C#/Labs-And-Homeworks/Entity Framework Core/EntityRelationsExercise/P03_FootballBetting.Models/Models/Player.cs
@@ -8,7 +8,7 @@
 
         public string Name { get; set; }
 
-        public int SquadNumber { get; set; }
+        public byte SquadNumber { get; set; }
 
         public int TeamId { get; set; }
